Added in waves and tested current game
Game will work as players can log in, set the volume, play the game, earn points, die and have their high score possibly updated, and return to the main menu. Neural network that resulted from over 2 hours of testing is not functioning properly. Unsure of the cause of this as there were no complications during testing and no errors, warnings, or comments during gameplay. Currently the ML Boss only chases the player and does not attack or block. Will look into this further.